Move ConflictingEnrollmentAlert
